+++ src/delineate_hydros_and_produce_HAND.sh
@@ -112,6 +112,30 @@ mpiexec -n $ncores_gw $taudemDir/gagewatershed \
     -o $tempCurrentBranchDataDir/demDerived_reaches_split_points_$current_branch_id.gpkg \
     -id $tempCurrentBranchDataDir/idFile_$current_branch_id.txt
 
+## POLYGONIZE REACH CATCHMENTS ##
+echo -e $startDiv"Polygonize Reach Catchments $hucNumber $current_branch_id"
+gdal_polygonize.py -q -8 -f GPKG $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.tif \
+    $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.gpkg catchments HydroID
+
+## DISSOLVE UNILATERAL CATCHMENTS AND REACHES ##
+if [ "$current_branch_id" != "$branch_zero_id" ] ; then
+    echo -e $startDiv"Dissolve Unilateral Catchments and Reaches $hucNumber $current_branch_id"
+    $srcDir/dissolve_unilateral_catchments.py \
+        -c $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.gpkg \
+        -r $tempCurrentBranchDataDir/demDerived_reaches_split_$current_branch_id.gpkg \
+        -p $tempCurrentBranchDataDir/demDerived_reaches_split_points_$current_branch_id.gpkg \
+        -co $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.gpkg \
+        -ro $tempCurrentBranchDataDir/demDerived_reaches_split_$current_branch_id.gpkg \
+        -po $tempCurrentBranchDataDir/demDerived_reaches_split_points_$current_branch_id.gpkg
+fi
+
+## RASTERIZE REACH CATCHMENTS ##
+echo -e $startDiv"Rasterize Reach Catchments $hucNumber $current_branch_id"
+gdal_rasterize -q -ot Int32 -a HydroID -a_nodata 0 -init 0 -co "COMPRESS=LZW" -co "BIGTIFF=YES" -co "TILED=YES" \
+    -te $xmin $ymin $xmax $ymax -ts $ncols $nrows \
+    $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.gpkg \
+    $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.tif
+
 ## VECTORIZE FEATURE ID CENTROIDS ##
 echo -e $startDiv"Vectorize Pixel Centroids $hucNumber $current_branch_id"
 $srcDir/reachID_grid_to_vector_points.py \
@@ -169,7 +193,7 @@ fi
 
 ## POLYGONIZE REACH WATERSHEDS ##
 echo -e $startDiv"Polygonize Reach Watersheds $hucNumber $current_branch_id"
-gdal_polygonize.py -q -8 -f GPKG $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.tif \
+gdal_polygonize.py -q -8 -overwrite -f GPKG $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.tif \
     $tempCurrentBranchDataDir/gw_catchments_reaches_$current_branch_id.gpkg catchments HydroID
 
 ## PROCESS CATCHMENTS AND MODEL STREAMS STEP 1 ##